+++ Graphique.hpp
@@ -10,14 +10,14 @@ public:
     Graphique();  // Constructeur
 
     // Méthode pour ajouter une courbe
-    void ajouterCourbe(const std::vector<double>& x, const std::vector<double>& y, const std::string& label, const std::string& couleur);
+    void ajouterCourbe(const std::vector<float>& x, const std::vector<float>& y, const std::string& label, const std::string& couleur);
 
     // Méthode pour afficher le graphique avec les courbes ajoutées
     void afficher();
 
 private:
-    std::vector<std::vector<double>> dataX; // Stockage des données X (temps)
-    std::vector<std::vector<double>> dataY; // Stockage des données Y (valeurs des variables)
+    std::vector<std::vector<float>> dataX; // Stockage des données X (temps)
+    std::vector<std::vector<float>> dataY; // Stockage des données Y (valeurs des variables)
     std::vector<std::string> labels;        // Noms des courbes
     std::vector<std::string> couleurs;      // Couleurs des courbes
 };
